feat(UI-1303): add redispatch functionality to events page
